Added ValueString to IProperty
Able to get string of value without knowing the property type.
It is generated for all values except EnumProperty, where it is set at construction. Ideally, all should be set at construction so it is 1:1 with the XML, but there are too many places to update, and generating the string is good enough.

Also in this commit: OverrideObject stack overflow bug fixed. Unsure what the original recursive structure was for.